(first iteration) application log here, juju logs still here, no log when success
